requirements.txt
streamlit
streamlit-option-menu
streamlit-lottie
streamlit-extras
streamlit-custom-notification-box
python-docx
pandas
plotly
pillow
pycryptodome==3.20.0
python-dotenv
pdfminer.six==20221105
nltk
scikit-learn
sqlalchemy
openpyxl
requests
spacy
pypdf==4.2.0
selenium
numpy
webdriver-manager
chromedriver-autoinstaller
google-generativeai
pdf2image
pytesseract
pdfplumber
reportlab
openrouter
docx2pdf
docx2txt
python-pptx
matplotlib
seaborn
pypdf2
language_tool_python==2.7.1
uvicorn
PyMuPDF
fastapi
python-multipart
pydantic
openai
# Phase 2 / speech packages
sounddevice==0.5.3
wavio==0.0.7
SpeechRecognition==3.14.3
CFFI==2.0.0
pycparser==2.23
soundfile==0.12.1

